polygen/requirements.txt
PyYAML
redis
requests
prometheus_client
